requirements.txt
feedparser
python-telegram-bot
python-dotenv
psycopg2-binary
pyrogram
TgCrypto
